# Conflicts:
#	logs/error/2024-11-22/error-0.log
#	logs/error/2024-11-25/error-0.log
#	logs/error/2024-11-26/error-0.log
#	logs/error/2024-11-27/error-0.log
#	logs/error/2024-11-28/error-0.log
#	logs/error/2024-12-02/error-0.log
#	logs/info/2024-11-22/info-0.log
#	logs/info/2024-11-25/info-0.log
#	logs/info/2024-11-26/info-0.log
#	logs/info/2024-11-27/info-0.log
#	logs/info/2024-11-28/info-0.log
#	logs/info/2024-11-29/info-0.log
#	logs/info/2024-12-02/info-0.log
#	logs/total/2024-11-22/total-0.log
#	logs/total/2024-11-25/total-0.log
#	logs/total/2024-11-26/total-0.log
#	logs/total/2024-11-27/total-0.log
#	logs/total/2024-11-28/total-0.log
#	logs/total/2024-11-29/total-0.log
#	logs/total/2024-12-02/total-0.log
#	logs/warn/2024-11-22/warn-0.log
#	logs/warn/2024-11-25/warn-0.log
#	logs/warn/2024-11-26/warn-0.log
#	logs/warn/2024-11-27/warn-0.log
#	logs/warn/2024-11-28/warn-0.log
#	src/main/resources/Mapper/ReportMapper.xml

+++ src/main/java/wh/fcfz/officecontroller/all/bean/Dao/form/FormTemplateFieldOption.java
@@ -19,12 +19,12 @@ public class FormTemplateFieldOption implements Serializable {
      * 表单模板列的可选项 ID
      */
     @TableId(type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     /**
      * 表单模板列 ID
      */
-    private Long templateFieldId;
+    private Integer templateFieldId;
 
     /**
      * 可选项值
